+++ packages/bunshine/src/middleware/headers/headers.spec.ts
@@ -4,12 +4,11 @@ import HttpRouter from '../../HttpRouter/HttpRouter';
 import { headers } from './headers';
 
 describe('headers middleware', () => {
-  let port = 50300;
   let server: Server;
   let app: HttpRouter;
   beforeEach(() => {
     app = new HttpRouter();
-    server = app.listen({ port: port++ });
+    server = app.listen({ port: 0 });
   });
   afterEach(() => {
     server.stop(true);