+++ packages/bunshine/src/middleware/performanceHeader/performanceHeader.spec.ts
@@ -4,12 +4,11 @@ import HttpRouter from '../../HttpRouter/HttpRouter';
 import { performanceHeader } from './performanceHeader';
 
 describe('performanceHeader middleware', () => {
-  let port = 50500;
   let app: HttpRouter;
   let server: Server;
   beforeEach(() => {
     app = new HttpRouter();
-    server = app.listen({ port: port++ });
+    server = app.listen({ port: 0 });
   });
   afterEach(() => {
     server.stop(true);